Added users DP file update feature using multer and got started with creating Own APIs with JWT authentication

+++ controllers/posts_controller.js
@@ -1,19 +1,48 @@
 const postsInfo = require('../models/posts_schema');
 const commentInfo = require('../models/comment_schema');
 
-module.exports.create_post = function(req,res){
-    postsInfo.create({
-        content:req.body.article,
-        user:req.user.id
-    },function(err,post){
-        if(err){
-            req.flash('error',err);
-            return res.redirect('back');
+module.exports.create_post = async function(req,res){
+    // postsInfo.create({
+    //     content:req.body.article,
+    //     user:req.user.id
+    // },function(err,post){
+    //     if(err){
+    //         req.flash('error',err);
+    //         return res.redirect('back');
+    //     }
+    //     // console.log(post);
+    // });
+    // req.flash('success','Post Created Successfully');
+    // return res.redirect('back');
+
+    try{
+        const post = await postsInfo.create({
+            content:req.body.article,
+            user:req.user.id
+        });
+
+        const postToSend = await postsInfo.findById(post._id).populate('user').populate({
+            path: 'comments',
+            populate:{
+                path : 'user_id'
+            }
+        });
+
+        if(req.xhr){
+            return res.status(200).json({
+                data:{
+                    post : postToSend
+                },
+                message: "Post created"
+            });
         }
-        // console.log(post);
-    });
-    req.flash('success','Post Created Successfully');
-    return res.redirect('back');
+
+        req.flash('error',err);
+        return res.redirect('back');
+    }catch(err){
+        req.flash('error',err);
+        return res.redirect('back');
+    }
 }
 
 module.exports.delete_post = async function(req,res){
@@ -35,12 +64,23 @@ module.exports.delete_post = async function(req,res){
 
     try{
         const post = await postsInfo.findById(req.params.id);
-        if(post.user == req.user.id){
+        if(post && post.user == req.user.id){
             const deleted = await commentInfo.deleteMany({"post_id":req.params.id});
             post.remove();
-            req.flash('success','Post Deleted');
+            if(req.xhr){
+
+                return res.status(200).json({
+                    data:{
+                        post_id : req.params.id
+                    },
+                    success: "Post Deleted"
+                });
+            }
+
+
             return res.redirect('back'); 
         }
+     
         req.flash('success','Your are Not authorized to delete this post');
         return res.redirect('back');
     }
